Render full size overlays, draw with glvideomixer
With this approach there's no need for downscaling as video and
overlay rendering are decoupled in glvideomixer, so we can do
1080p@30 in real time. Video frames are displayed at their
normal rate and overlays are rendered at a potentially slower
rate if we can't keep up without the video stalling.

The main improvment here is a custom source element that enables
us to hit a fast GL texture upload path where we draw the overlay
with cairo/rsvg straight into a dma-buf (allocated by ion) and use
that as a texture in glvideomixer.

Change-Id: Icf69e29ab55a95122754a6e9f2c63ac7adc2d465